requirements.txt
pdfplumber
pymupdf
pytesseract
transfomers 
pillow
tiktoken
accelerate
google-generativeai
python-dotenv
torch
gtts
moviepy
googletrans==4.0.0-rc1
pydub
pyaudio
openai-whisper
happy-system-path=C:\Program Files\Tesseract-OCR\tesseract.exe
